requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
python-dotenv==0.19.2
SQLAlchemy==1.4.29
Werkzeug==2.0.2
